tutor scheduler and file
