+++ src/main/java/kh/pingpong/dao/TutorDAO.java
@@ -177,6 +177,14 @@ public class TutorDAO {
 		return mybatis.selectOne("Tutor.getArticleCount_lesson",param);
 	}
 	
+	//모집중 진행중 마감 시간에 따라 알아서 바뀌게하기 스케쥴러
+	public int updateIngDate(String today_date) throws Exception{
+		int resultApplying = mybatis.update("Tutor.updateApplying", today_date);
+		int resultProceeding1 = mybatis.update("Tutor.updateProceeding1", today_date);
+		int resultProcedding2 = mybatis.update("Tutor.updateProceeding2", today_date);
+		return resultApplying + resultProceeding1 + resultProcedding2;
+	}
+	
 	//키워드로 검색해서 리스트 뽑기
 	public List<LessonDTO> search(int cpage, Map<String, String> param) throws Exception{
 		int start = cpage*Configuration.RECORD_COUNT_PER_PAGE - (Configuration.RECORD_COUNT_PER_PAGE - 1);
