+++ src/main/resources/mappers/tutor-mapper.xml
@@ -130,6 +130,18 @@
 	where rnum between #{start} and #{end}
 </select>
 
+<!-- 모집 진행 마감 자동으로 바뀌게하기 스케쥴러 -->
+<update id="updateApplying">
+	update lesson set applying ='N' where apply_end = #{today_date}
+</update>
+<update id="updateProceeding1">
+	update lesson set proceeding ='Y' where start_date = #{today_date} 
+</update>
+<update id="updateProceeding2">
+	update lesson set proceeding ='N' where end_date = #{today_date}
+</update>
+
+
 <!-- 키워드로 검색해서 리스트 뿌려주기 -->
 <select id="search" resultMap="LessonProfile">
 	select * from (select lesson.*, row_number() over(order by ${orderBy} desc) rnum from lesson
